made props more consistent and file restructure
