--- frog/imports/client/UserDashboard/TopBarWrapper/TopBarWrapper.js
@@ -1,7 +1,0 @@
-// @flow
-import * as React from 'react';
-import { TopBarAccountsWrapper } from './TopBarAccountsWrapper';
-
-export const TopBarWrapper = ({ title }) => {
-  return <TopBarAccountsWrapper title={title} />;
-};
