--- frog/imports/client/UserDashboard/TopBarWrapper/types.js
@@ -1,9 +1,0 @@
-export type MenuItemT = {
-  name: string,
-  icon: React.Component<*>,
-  callback: () => void
-};
-export type TopBarWrapperPropsT = {
-  title: string,
-  menuItems: Array<MenuItemT>
-};
