+++ frog/imports/containers/TopBarWrapper/types.js
@@ -1,0 +1,4 @@
+export type TopBarWrapperPropsT = {
+  navigation: React.Element<*>,
+  actions: React.Element<*>
+};
